move skydriver communication up the call stack
